refactor(widget): make the route-issue card testable
`buildCard` was private to the hook, so the logic that decides every suggestion
— which figure to offer, and whether to offer a button at all — had no test at
all. It is a pure function of an issue and a handful of values; only the
gathering of those values needs React.

Moved it to `utils/routeIssues/card.ts` as `buildRouteIssueCard`, leaving the
hook to assemble the dependencies. No behaviour change.

The new test covers what the reports in this PR were about: a suggestion that
would not move the amount offers no button, a slippage cap withdraws its button
once applied, a cap offers nothing on a resolved auto value, and a contract-call
quote with no send amount invents no figure.
